[refactor](jdbc catalog) refactor jdbc catalog get databases logic
