+++ fe/fe-core/src/main/java/org/apache/doris/datasource/jdbc/client/JdbcOracleClient.java
@@ -38,39 +38,8 @@ public class JdbcOracleClient extends JdbcClient {
     }
 
     @Override
-    protected String getDatabaseQuery() {
-        return "SELECT DISTINCT OWNER FROM all_tables";
-    }
-
-    @Override
-    protected String getCatalogName(Connection conn) throws SQLException {
-        return conn.getCatalog();
-    }
-
-    @Override
     public String getTestQuery() {
         return "SELECT 1 FROM dual";
-    }
-
-    @Override
-    public List<String> getDatabaseNameList() {
-        Connection conn = getConnection();
-        ResultSet rs = null;
-        if (isOnlySpecifiedDatabase && includeDatabaseMap.isEmpty() && excludeDatabaseMap.isEmpty()) {
-            return getSpecifiedDatabase(conn);
-        }
-        List<String> remoteDatabaseNames = Lists.newArrayList();
-        try {
-            rs = conn.getMetaData().getSchemas(conn.getCatalog(), null);
-            while (rs.next()) {
-                remoteDatabaseNames.add(rs.getString("TABLE_SCHEM"));
-            }
-        } catch (SQLException e) {
-            throw new JdbcClientException("failed to get database name list from jdbc", e);
-        } finally {
-            close(rs, conn);
-        }
-        return filterDatabaseNames(remoteDatabaseNames);
     }
 
     @Override
