+++ regression-test/data/external_table_p0/jdbc/test_mariadb_jdbc_catalog.out
@@ -2,6 +2,11 @@
 -- !sql --
 internal
 
+-- !show_db --
+doris_test
+information_schema
+performance_schema
+
 -- !sql --
 internal
