+++ regression-test/data/external_table_p0/jdbc/test_oracle_jdbc_catalog.out
@@ -1,4 +1,12 @@
 -- This file is automatically generated. You should know what you did if you want to edit this
+-- !show_db --
+ANONYMOUS
+APEX_040000
+APEX_PUBLIC_USER
+DORIS_TEST
+HR
+information_schema
+
 -- !test0 --
 1	alice	20	99.5
 2	bob	21	90.5
