+++ regression-test/data/external_table_p0/jdbc/test_pg_jdbc_catalog.out
@@ -1,4 +1,11 @@
 -- This file is automatically generated. You should know what you did if you want to edit this
+-- !show_db --
+catalog_pg_test
+doris_test
+information_schema
+pg_catalog
+public
+
 -- !test0 --
 123	abc
 123	abc
@@ -2276,7 +2283,6 @@ information_schema
 catalog_pg_test
 information_schema
 pg_catalog
-pg_toast
 public
 
 -- !specified_database_4 --
