+++ regression-test/data/external_table_p0/jdbc/test_sqlserver_jdbc_catalog.out
@@ -1,4 +1,19 @@
 -- This file is automatically generated. You should know what you did if you want to edit this
+-- !show_db --
+db_accessadmin
+db_backupoperator
+db_datareader
+db_datawriter
+db_ddladmin
+db_denydatareader
+db_denydatawriter
+db_owner
+db_securityadmin
+dbo
+guest
+information_schema
+sys
+
 -- !test0 --
 1	doris	18
 2	alice	19
